python_core/requirements.txt
beautifulsoup4
emoji
fonttools
tinycss2
tqdm
pillow
ttkbootstrap
tkinterdnd2
pypinyin
opencc-python-reimplemented
